% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GOFs.R
\name{add.caption}
\alias{add.caption}
\title{Add caption to ggplot object}
\usage{
add.caption(p, control = GOF.control())
}
\arguments{
\item{p}{ggplot object}

\item{control}{an optional list of control settings. See GOF.control for the names of the settable control values and their effect.}
}
\description{
Adds caption text as returned by get.caption to a ggplot object
}
