% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{extFormat}
\alias{extFormat}
\title{Internal package function}
\usage{
extFormat(ext, format.estimate = "\% -#6.4g", format.rse = "\%#6.3g")
}
\arguments{
\item{ext}{the result ofname of the lst file without the .lst extension. model may include full or relative path to the lst file.}

\item{format.estimate}{format for estimated value, passed to sprintf}

\item{format.rse}{format for rse, passed to sprintf}
}
\value{
a character-matrix
}
\description{
Formatting of the named list returned by extload
}
\keyword{internal}

