% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{extload.sub.table}
\alias{extload.sub.table}
\title{Internal package function}
\usage{
extload.sub.table(
  model,
  skip,
  nrows,
  positive.iterations.only,
  use.model.path = TRUE
)
}
\arguments{
\item{model}{model}

\item{skip}{number of rows to skip}

\item{nrows}{number of rows to read}

\item{positive.iterations.only}{positive iterations only}

\item{use.model.path}{Load file from a global defined model library (TRUE=default).
If so will look for a global character vector named \code{model.path}}
}
\value{
Named list including theta, omega, sigma, and ofv. For MCMC output file each object are matrixes.
}
\description{
Internal package function
}
\keyword{internal}
