% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{extTransform}
\alias{extTransform}
\title{Internal package function}
\usage{
extTransform(ext, type = 2)
}
\arguments{
\item{ext}{named list according to output from extload}

\item{type}{Type=0: Present OMEGA and SIGMA as variance and covariances and display SE for THETA, OMEGA, SIGMA
Type=1: Present OMEGA and SIGMA as variance and covariances and display RSE for THETA, OMEGA, SIGMA
Type=2: Present OMEGA and SIGMA as standard-deviation and correlations and display RSE for THETA, OMEGA, SIGMA
Type=3: Present OMEGA and SIGMA as standard-deviation and correlations and display SE for THETA, OMEGA, SIGMA}
}
\value{
named list according to output from extload
}
\description{
Transforming of the named list returned by extload
}
\keyword{internal}

