% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.control.R
\name{get.label}
\alias{get.label}
\title{Labels for known NONMEM variables}
\usage{
get.label(x, trans = NULL)
}
\arguments{
\item{x}{column to get label for}

\item{trans}{transformation}
}
\description{
get.label match known NONMEM variables to the GOF-dictionary and returns the matched label.
Unless trans is NULL, the label is modified to 'f(matched label)'
}
\seealso{
[get.GOF.dictionary()],  [set.GOF.dictionary()], and [default.GOF.dictionary()].
}
