% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.control.R
\name{get.model.path}
\alias{get.model.path}
\alias{set.model.path}
\title{Get and set model.path}
\usage{
get.model.path()

set.model.path(model.path)
}
\arguments{
\item{model.path}{character string for path to model files}
}
\description{
The model.path in the nonmem2R set of global variables is used by functions
loading NONMEM or PSN generated output file. The model.path can ge retrived by get.model.path and
can be set using set-model.path. When loading nonmem2R the model-path is set to "",
e.g. the current working directory. The model.path can be set abolute or relative, see examples.
}
\examples{
\dontrun{
## Example for setting absolute path
set.model.path("c:/NONMEM")
## Example for setting relative path,
set.model.path("../NONMEM")
}
}
