% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{grid.sim.internal2}
\alias{grid.sim.internal2}
\title{Internal function}
\usage{
grid.sim.internal2(n, means = NULL, sigma, grid.param = NULL)
}
\arguments{
\item{n}{Number of simulations}

\item{means}{vector of mean values}

\item{sigma}{covariance matrix}

\item{grid.param}{the index of the parameter for which a fix grid (from qnorm) is used instead of a simulated grid (rnorm).
If grid.param=NULL ( default) the fix grid will be used for the parameter with largest variance.}
}
\value{
row-matrix of parameters
}
\description{
Internal function
}
\keyword{internal}

