% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{grid.sim.internal3}
\alias{grid.sim.internal3}
\title{Internal function}
\usage{
grid.sim.internal3(n, means = NULL, sigma)
}
\arguments{
\item{n}{Number of simulations}

\item{means}{vector of mean values}

\item{sigma}{covariance matrix}
}
\value{
row-matrix of parameters
}
\description{
Internal function
}
\keyword{internal}
