% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_GOFs.R
\name{merge2GOF}
\alias{merge2GOF}
\alias{merge4GOF}
\alias{merge6GOF}
\title{Merging 2, 4, or 6 GOF's into one graph}
\usage{
merge2GOF(p1, p2, byrow = TRUE)

merge4GOF(p1, p2, p3, p4)

merge6GOF(p1, p2, p3, p4, p5, p6)
}
\arguments{
\item{p1}{first ggplot object}

\item{p2}{second ggplot object}

\item{byrow}{combine plots side by side (TRUE) or one above the other (FALSE)}

\item{p3}{third ggplot object}

\item{p4}{4th ggplot object}

\item{p5}{5th ggplot object}

\item{p6}{6th ggplot object}
}
\description{
Merging multiple GOFs (ggplot objects) into one graph with
caption added as reurned by get.caption. Any legend present in first GOF is added to the combined graph.
}
