% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{modload}
\alias{modload}
\title{Show model file}
\usage{
modload(model, use.model.path = TRUE)
}
\arguments{
\item{model}{name of the mod file with or without the .mod extension.
model may include full or relative path to the mod file. See examples.}

\item{use.model.path}{Load file from a global defined model library (TRUE=default).
If so will look for a global character vector named \code{model.path}}
}
\value{
data.frame
}
\description{
Load and dump model file in the consol.
}
\examples{
##### Load the .ext file "run001.mod"
# 1) Get path to the example file included in nonmem2R package
file1 <- system.file("extdata", "run001.mod", package = "nonmem2R")
# 2) Load the file using the extload function
modload(file1)

}
