% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\name{position_scale}
\alias{position_scale}
\title{Position for scaling y}
\usage{
position_scale(y = 1)
}
\arguments{
\item{y}{scale value}
}
\description{
Position for scaling y
}
\details{
Position for scaling y
}
