% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{PositionScale}
\alias{PositionScale}
\title{ggproto for position_scale}
\format{An object of class \code{PositionScale} (inherits from \code{PositionDodge}, \code{Position}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
PositionScale
}
\description{
ggproto for position_scale
}
\keyword{internal}
