% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{print.extToTable}
\alias{print.extToTable}
\title{Print function for compiled summary information of class extToTable}
\usage{
\method{print}{extToTable}(x, ...)
}
\arguments{
\item{x}{a object af class extToTable.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Print function for compiled summary information of class extToTable
}
\keyword{internal}

