% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{print.modload}
\alias{print.modload}
\title{Print function for loaded mod file}
\usage{
\method{print}{modload}(x, ...)
}
\arguments{
\item{x}{a object af class modload.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Print function for loaded mod file class modload
}
\keyword{internal}
