% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{print.sumoR}
\alias{print.sumoR}
\title{Print function for compiled summary information of class sumoR}
\usage{
\method{print}{sumoR}(x, ...)
}
\arguments{
\item{x}{a object af class sumoR.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Print function for compiled summary information of class sumoR
}
\keyword{internal}

