% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_read.R
\name{readVpc}
\alias{readVpc}
\title{Internal package function}
\usage{
readVpc(vpc.results = NULL, npc.results = NULL, verbose = FALSE, ...)
}
\arguments{
\item{vpc.results}{vpc results}

\item{npc.results}{npc results}

\item{verbose}{verbose}

\item{...}{Further arguments}
}
\value{
Named list
}
\description{
Internal package function
}
\keyword{internal}
