% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\name{stat_dnorm}
\alias{stat_dnorm}
\title{Normal density}
\usage{
stat_dnorm(mapping = NULL, data = NULL, geom = "line",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{geom}{Use to override the default geom}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{other arguments passed on to layer.}
}
\description{
Normal density with ggplot2
}
\details{
Provide normal QQplot
}
