% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{StatCorr}
\alias{StatCorr}
\title{ggproto for stat_corr}
\format{An object of class \code{StatCorr} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatCorr
}
\description{
ggproto for stat_corr adding correlation to x-y plots
}
\keyword{internal}
