% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{StatDenx}
\alias{StatDenx}
\title{ggproto for stat_denx}
\format{
An object of class \code{StatDenx} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatDenx
}
\description{
ggproto for stat_denx adding density for x with y-scaled to fit current scale
}
\keyword{internal}
