% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{StatDnorm}
\alias{StatDnorm}
\title{ggproto for stat_dnorm}
\format{An object of class \code{StatDnorm} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatDnorm
}
\description{
ggproto for stat_dnorm doing normal density to histograms
}
\keyword{internal}
