% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{StatHmean}
\alias{StatHmean}
\title{ggproto for stat_hmean}
\format{
An object of class \code{StatHmean} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatHmean
}
\description{
ggproto for stat_hmean doing horizontal reference line at mean
}
\keyword{internal}
