% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_and_stats.R
\docType{data}
\name{StatHmedian}
\alias{StatHmedian}
\title{ggproto for stat_hmedian}
\format{
An object of class \code{StatHmedian} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatHmedian
}
\description{
ggproto for stat_hmedian doing horizontal reference line at median
}
\keyword{internal}
