% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{sumoR}
\alias{sumoR}
\title{Compile summary information of for NONMEM model based on the lst file, ext file, and the cov file.}
\usage{
sumoR(model)
}
\arguments{
\item{model}{name of the lst file without the .lst extension. model may include full or relative path to the lst file.}
}
\value{
named list of class sumoR
}
\description{
Compile summary information similar to that of the sumo PSN function, based on the NONMEM output files
lst, ext, and if covariance setp was run, the cov file.
}

