% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{sumoRU}
\alias{sumoRU}
\title{Internal package function}
\usage{
sumoRU(file.path)
}
\arguments{
\item{file.path}{path including file name to model lst file}
}
\value{
named list
}
\description{
Internal function for sumoR on unix/linux system
}
\keyword{internal}

