% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{test.grid.sim}
\alias{test.grid.sim}
\title{Function for testing grid.sim and compare with rmvnorm}
\usage{
test.grid.sim(n = 1000, k = 1:4)
}
\arguments{
\item{n}{number of simulations}

\item{k}{subset of parameters from a 4X4 sigma to use}
}
\value{
graphics
}
\description{
Test grid.sim
}
\examples{
\dontrun{
require(lattice)
test.grid.sim(n=1000)
}
}
