% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNonmemOutput.R
\name{systemPSN}
\alias{systemPSN}
\title{Run qpsn system call}
\usage{
systemPSN(cmd, use.model.path = TRUE, ml = "ml psn nonmem-standard",
  ...)
}
\arguments{
\item{cmd}{qpsn cmd}

\item{use.model.path}{Run cmd in a specified model library (TRUE=default).
If so will look for a global character vector named \code{model.path} and
run the system cmd the folder path as specified in \code{model.path}.}

\item{ml}{String for what module to load}

\item{...}{Further arguments}
}
\description{
Run qpsn system call on system where psn is available.
If psn is not available system{base} is used.
}
\examples{
# List files in working directory
try(systemPSN("ls -l"))

# List file in working directory indluding sun-directories
try(systemPSN("ls * -l"))
}
